requirements.txt
blinker==1.8.2
Cerberus==1.3.5
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
looseversion==1.3.0
MarkupSafe==3.0.1
python-dotenv==1.0.1
PyYAML==6.0.2
rethinkdb==2.4.10.post1
six==1.16.0
Werkzeug==3.0.4
